add greetings after 10 page loads add opportunity to get data from json generate schema info via anotate gem

+++ db/migrate/20161117104724_add_user_id_to_post.rb
@@ -1,5 +1,5 @@
 class AddUserIdToPost < ActiveRecord::Migration[5.0]
   def change
-    add_column :posts, :user_id, :decimal
+    add_column :posts, :user_id, :integer
   end
 end
